requirements/base.txt
amqp==5.0.2
asgiref==3.3.1
attrs==20.3.0
billiard==3.6.3.0
cairocffi==1.2.0
CairoSVG==2.5.1
celery==5.0.5
cffi==1.14.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
cssselect2==0.4.1
cycler==0.10.0
defusedxml==0.6.0
Django==3.1.4
django-crispy-forms==1.10.0
django-upload-validator==1.1.5
html5lib==1.1
importlib-metadata==3.3.0
iniconfig==1.1.1
Jinja2==2.11.2
kiwisolver==1.3.1
kombu==5.0.2
MarkupSafe==1.1.1
matplotlib==3.3.3
numpy==1.23.3
packaging==20.8
pandas==1.4.4
Pillow==8.1.0
pluggy==0.13.1
prompt-toolkit==3.0.8
py==1.10.0
pycparser==2.20
pyparsing==2.4.7
Pyphen==0.10.0
pytest==6.2.1
python-dateutil==2.8.1
python-decouple==3.3
python-magic==0.4.18
pytz==2020.5
scipy==1.9.1
six==1.15.0
sqlparse==0.4.1
tinycss2==1.1.0
toml==0.10.2
typing-extensions==3.7.4.3
tzlocal==2.1
vine==5.0.0
wcwidth==0.2.5
WeasyPrint==52.2
webencodings==0.5.1
zipp==3.4.0
dj-database-url==0.5.0
psycopg2-binary==2.9.1
scikit-learn==1.1.2
